requirements.txt
python-dotenv
discord-py-interactions
jurigged
rich
GitPython